Revert "Merge branch 'transitionECS' into Dev4"
This reverts commit ea90f6f2566ce412f42d71ef8fb6991d42eb22af, reversing
changes made to 83864894841094b06301e77b0519ec9550d17ff5.

+++ Project-X/Tool/Animation.h
@@ -13,7 +13,7 @@ private:
     
     sf::Vector2i sizeSpriteSheet;
     int nbrFrames;
-    float frameRate;
+    int frameRate;
     
     sf::Vector2i locTexture;
     int actualFrame;
